Removed unmerged .pyc files and resolved merge conflict

+++ base/views.py
@@ -1,24 +1,85 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
-# Create your views here.
+from django.contrib.auth import authenticate, login, logout
 from .models import Room,Topic
 from .forms import RoomForm
+from django.contrib.auth.forms import UserCreationForm
+from django.db.models import Q
+from django.contrib.auth.models import User
 
 
-rooms = [
-    {'id': 1, 'name': 'lets learn python'},
-    {'id': 2, 'name': 'Design with me'},
-    {'id': 3, 'name': 'Front end developers'},
-]
+# rooms = [
+#     {'id': 1, 'name': 'lets learn python'},
+#     {'id': 2, 'name': 'Design with me'},
+#     {'id': 3, 'name': 'Front end developers'},
+# ]
 
+def loginpage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+
+
+    if request.method =='POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist.")
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username or password does not exist.")
+
+           
+            
+    context = {'page':page}
+    return render(request, 'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request,'an error occured during registration')
+
+    return render(request, 'base/login_register.html',{'form':form})
 
 
 def home(request):
     q = request.GET.get('q') if  request.GET.get('q') != None else ''
-    rooms = Room.objects.filter(topic__name__icontains=q)
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q)|
+        Q(name__icontains=q) |
+        Q(description__icontains=q) |
+        Q(host__username__icontains=q) 
+        )
     topic = Topic.objects.all()
-
-    context = {'rooms': rooms , 'topics': topic}
+    room_count = rooms.count()
+    context = {'rooms': rooms , 'topics': topic , 'room_count': room_count} 
 
     return render(request, 'base/home.html', context)
 
@@ -31,6 +92,8 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -42,10 +105,13 @@ def createRoom(request):
 
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to update this room')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -55,8 +121,14 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def  deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to delete this room')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
